requirements.txt
django==4.0.4
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pillow==9.1.1

